requirements.txt
alembic==1.3.1
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
gunicorn==19.7.1
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.1
Mako==1.1.0
MarkupSafe==1.1.1
numpy==1.18.0
Pillow==7.0.0
psycopg2==2.7.3.1
python-dateutil==2.8.1
python-editor==1.0.4
scikit-learn==0.22
scipy==1.4.1
six==1.13.0
sklearn==0.0
SQLAlchemy==1.3.11
tflearn==0.3.2
Werkzeug==0.16.0